Do not use test specific db's in mongoid config
Realised when attempting to get the feature tests passing after having updated the project to use MongoDb 3.6 that we were getting errors because the front end and the waste-carriers-service were out of sync.

The feature tests rely on the waste-carriers-service running, but without having 2 versions running on different ports and pointing at different databases there would be no way to run them using a test specific database.

Therefore this change updates the `mongoid.yml` to remove references to a test specific registrations and users database when using the test client.
 2 versions running on different ports and pointing at different databases there would be no way to run them using a test specific database.

Therefore this change updates the `mongoid.yml` to remove references to a test specific registrations and users database when using the test client.