Fix format of payment dateReceived in test helper
With the update to the waste-carrier-service and the Java dependencies it uses to support working with MongoDb 3.6, it would appear it has become more strict about the format of the Json it will accept.

Before this change calls to the add a payment endpoint had started returning a http 400 bad request. Re-formatting the way the date is formatted to pad the month and day with zeroes appears to resolve the issue.