Update mongoid behaviour. No error when not found
Highlighted when trying to seed the database, our logic currently assumes functions like `find_by_email()` will return null when nothing is found.

However with the latest mongoid driver the default is actually to throw an error. By amending the configuration we can get back to a behaviour we expect.