Upgrade service to 2.4.2 to match wcr renewals
https://eaflood.atlassian.net/browse/WC-234

The waste carriers renewals project is currently using 2.4.2 whilst this was using 2.3.1. Locally we deploy them onto one box and we anticipate doing so in our new AWS environments. Therefore to simplify maintenance and deployment issues we are upgrading this project to a matching version.

This means wherever we deploy the apps to in the future now only has to worry about one version of Ruby.