Fix README details and defaults for services
https://eaflood.atlassian.net/browse/WC-157
https://eaflood.atlassian.net/browse/WC-159

After updating the Waste carriers service and OS Places address lookup projects the defaults for them have changed. This change covers updating the defaults and info relating to them in the README.

Whilst in the README also took the time to clear up the other information it contains to better reflect the current state of the service and project.