Update ohm and rails versions
Updated the ohm gem to version 3.0.0 because after updating to use 2.4.2 and running bundle install we faced the following error

```
An error occurred while installing msgpack (0.5.12), and Bundler cannot continue.
Make sure that `gem install msgpack -v '0.5.12'` succeeds before bundling.

In Gemfile:
  ohm was resolved to 2.3.0, which depends on
    msgpack
```

The next (and latest) version of Ohm available is 3.0.0 so we updated the project to use that. This resolves the error mainly due to one of the 2 changes in this version is replacing msgpack for JSON. https://github.com/soveran/ohm/releases/tag/3.0.0

Also updated the version of rails to resolve the following error we were receiving when running the unit tests

```
warning: constant ::Fixnum is deprecated
warning: constant ::Bignum is deprecated
```

In ruby 2.4 Fixnum and Bignum were combined into a new type Integer. Getting a warning shouldn't break anything supposedly, but in our case it does. Upgrading the version of Rails used (and therefore the version of ActiveSupport where the stack trace points to these methods being used) resolves the issue. We decided to upgrade to the same version as is currently being used in the Waste carriers renewals service.