Handle nil values in call to Money.new
Following the upgrade to 2.4.2 running the feature tests led to a number of failed tests, with the same error reported in all cases

```
Money.new(some_value)
ArgumentError: invalid value for BigDecimal(): ""
```

Some searching (https://github.com/kubo/ruby-oci8/issues/143#issuecomment-266873632) lead to a reported inconsistency with how BigDecimal handles invalid values compared to others like Float and Integer. If passed an invalid error it still seems to return a value rather than throw an error.

Or used to till we updated the version of ruby and some core libraries. We think our invalid value in these cases is null, and amending any call to `Money.new()` to handle this has resolved the failing tests.

If you look through the changes here you'll see we've applied it where we pass in a value from an instance or when called within a method. If we are passing in a config value we don't handle nil with the expectation this should throw an error as the value should have been set.