Remove use of thin web server from feature tests
When using the thin web server issues with the deprecated use of Fixnum kept apprearing in the console output, as well as a message about the deprecated use of passing a block to Capybara::server.

As there is no real need to use thin when running the feature tests, this change removes it (more specfically the file that enables it to be used).