Revert capybara and capybara-email upgrades
When running the feature tests for some scenarios we were getting

```
Request failed to reach server, check DNS and/or server status (Capybara::Poltergeist::StatusFailError)
```

when we attempted to click on the verification link in an email. There didn't appear to be anything wrong with the link, and changing the env vars to use 0.0.0.0 or 127.0.0.1 made no difference.

We then tried reverting the versions of these gems, and that got them passing again.

At this point with the project in this state with the gems reverted all unit tests and all feature tests are passing.