Remove commented out reference to Nokogiri
We don't need to specify nokogiri in the gemfile. Every gem and his dog brings this in, so there is no need to specify it, unless you are trying to pin it down to a specific version.

We're not, but we commented it out whilst we confirmed this fact. Having confirmed it, we really should have yanked this line.