Synch secret key base & devise init. with renewals
https://eaflood.atlassian.net/browse/WC-306

In order to allow a user to move between the two systems (renewals and frontend) we need to ensure that devise is using the same `SECRET_KEY_BASE` across both.

The frontend was specifically using one env var to initialise Devise, and another in the app (for setting secure cookies etc), none of which were using the `secrets.yml` method in later services.

This change essentially copies the `secrets.yml` from renewals service, and removes the ad-hoc processes it was using and lets Rails use its default methods for handling the `SECRET_KEY_BASE`.