Make cookie serializer compatible with renewals
In order for the 2 services to be able to read the same cookie (which is the key to moving between the two) we also need to ensure they are serializing it in the same way.

Prior to this change the frontend was still using Marshal which is the default for rails apps prior to 4.1. After this a `config/intializers/cookies_serializer.rb` is added that specifies JSON.

As this was never added when the frontend was updated to 4.2.6 two years ago we need to add it now.

Final point is we specify `:hybrid` rather than `:json` as this setting is specifically to support apps which need to handle users with cookies that have not yet been converted. Failure to do so causes a serverside error as soon as someone with a Marshal serialized cookie visits the page.