Fix broken rake by upgrading Rspec
https://eaflood.atlassian.net/browse/WC-234

Following our upgrade to some of the projects dependencies rake has been upgraded to 12.3.1.

However Rake 11 removes a method that rspec-core < 3.4.4 depends on (https://makandracards.com/makandra/43932-error-undefined-method-last_comment)

The quick and easy fix for this would be pinning Rake to a lesser version, but the better solution is to upgrade the version of Rspec. We've gone with upgrading Rspec (to at least match the version in use by the renewals service).

So in this change we up upgraded RSpec-rails to 3.7.2 (~> 3.6). This then caused a dependency error with Capybara and subsequently capybara-email. We had to upgrade to the versions in this commit because earlier ones caused the following deprecation note to appear in the output

```
Capybara::RSpecMatchers::HaveText implements a legacy RSpec matcher
protocol. For the current protocol you should expose the failure messages
via the failure_message and failure_message_when_negated methods.
```

Specifically some searching online found the cause was a change in the naming of some methods in rspec 3 which capybara relies on. Hence the only fix was to update it (along with capybara-email) to a later version.