Update endpoints for various calls to WCR services
https://eaflood.atlassian.net/browse/WC-309

After making changes and a number of refactorings to the waste-carriers-service project the changes included here update the frontend to match.

The changes were all related to removing the dependency on Elasticsearch in the project, and as a consequence some housekeeping and refactoring of the query logic.

This has led to a change in some of the http calls made by the frontend hence the changes included here.

WIP