Fix display & export of regs with no keypeople
With the update to the waste carriers service's dropwizard and mongodb framework it now honours anything with a `@JsonProperty` when saving to the database.

This means any historic records where the fields are named as per the POJO are loosing data when the service's reads the record from MongoDb.

In our case this impacts the exports because they have logic which assumes that anything that isn't lower tier will have at least one key person. So we can at least see the registrations found these changes update the logic to handle this situation.