Convert payment status from array to single value
Allowing users to select any option of Fully paid, overpaid, and underpaid is meaningless. It really should just be one of the options.

Therefore we have updated the waste carriers service search such that this is now a required field, and we only expect users to select one of them.

This change updates the payments report functionality to match this, even including a validation to ensure it is selected.