Update renewals holding page with latest details
https://eaflood.atlassian.net/browse/WC-373

With the need to use the extensions and the shift to a delivery date of August 8 2018 it was felt by the team that we really need to update the renewals holding page (the page displayed when a user enters a registration number that is eligible for renewal) to reflect this.

This changes the month we aim to deliver from April to August, and based on the latest RPS removes any mention of how long registrations will be extended for.

WIP