Remove public search
The last vestigial of Elasticsearch is the expectation that the public register functionality in the frontend has that its in use. Hence the query it sends to the Waste carriers service.

As it's not, and the public register is defunct, the endpoint in the waste carriers service has been removed. This means the features tests for the public search will no longer pass, nor will the functionality work.

So this changes removes it from the project rather than leave non-working code, or just commenting it out.