Remove references and dependencies on elasticsearch (#134)
https://eaflood.atlassian.net/browse/WC-342

With the migration of Convictions service into the waste-carriers-service we have removed the last thing that properly uses elasticsearch.

This means we can now also remove any reference to its use here, which mainly seems to be in support of testing.

We also remove any functionality that depends on Elasticsearch being used by others, namely the public register features.

This last vestigial of Elasticsearch is the expectation that the public register functionality in the frontend believes its in use. Hence the query it sends to the Waste carriers service.

As it's not, and the public register is defunct, the endpoint in the waste carriers service has been removed. This means the feature tests for the public search and any steps will no longer pass, nor will the functionality work.

So this change also removes it from the project rather than leave non-working code, or just commenting it out.