Update url for address lookup
https://eaflood.atlassian.net/browse/WC-366

WC-366 is a story about adding the ability to 'mock' interactions with Worldpay in preparation for performance testing. As part of that story some general refactoring of the OS Places address lookup project were undertaken, including removing the need to specify `.json` in the url.

This change updates the frontend accordingly.