Fix broken conviction check after reg edit
https://eaflood.atlassian.net/browse/WC-342

Testing exposed an issue with the convictions check. Creating a new registration isn't an issue and it is correctly flagged. However if you then edit the registration and cause a change that forces a new one, the conviction check o the second pass would error.

After some investigation we were able to pin down the error to

- FE queries services for registration
- Service returns a JSON hash where all the dates are in milliseconds since epoch
- FE initialises registration based on hash
- FE when init the key people correctly converts milliseconds to date, but does not assign it to key_person.dob. Instead it assigns the milliseconds value
- FE then runs conviction check passing services milliseconds date
- Service errors as it is expecting to convert strings like '1981-01-01' into date values

The reason this is now failing as that the current production version of the frontend talks to the old convictions service, and never passed the date of birth to it. Its only because we are now properly using DOB in the conviction check that we are the services to interpret what the front end is sending.

We did initially try just setting key_person.dob to what ApplicationController.helpers.convert_date passes back. However this caused a nunmber of cucumber feature tests to fail.

So instead we just ensure that what we send to the services is what it is expecting by adding a new date formatter to the Key_person class.