Standardise comp. house config inc. host addr
This change is part of standardising the env var across all projects that holds the service's Companies House API key.

It also includes a change so that in future the host address for Companies house can be read from an env var (though we do default it if the env var is not set). This allows us in the future to point the app to a custom url, for example when mocking Companies House during a load test.