Standardise Frontend domain address
This changes updates the name of the env var that holds the address of the Frontend service, making it consistent across all projects.

The previous version of the variable used to be held as `localhost:3000` (for example). However the other apps using this hold the value as `http://localhost:3000`. It seems the Frontend relies on the version without the protocol for its email functionality.

So rather than create what in essence is a duplicate environment variable with the same value, we have also added some logic to strip the protocol from the url provided, and set that to a new config property. We then use that where the code previously was reading straight from the environment variable.