Standardise subdomain addresses
This changes updates the names of the env vars that holds the subdomain address of the frontend (e.g. localhost:3000), making it simpler to manage.

On seeing we had this functionality we also strip out some changes made in an earlier commit, as we no longer extrac t what we need from another env var.