Standardise use XVFB for PDF env var
This change updates the name of the env var that holds whether we should be using XVFB for generating PDF's. Essentially this is switched on in production, and when working in the vagrant box. Outside that and its up to the user to install and setup the ability to generate PDF's.