Make Worldpay url configurable and standardise it
This change makes it possible to change the worldpay url via an env var. This will allow us to set the url to point to production worldpay in our production environment, as well as allow us to set it to custom values e.g. point to a mock during load testing.

The choice of the env var is also standardised across the projects.