Remove default values for registration charges
Having tried not setting these env vars and expecting the defaults to kick in, we realised that we had misunderstood how a command would work.

If an env var does not exist the call `ENV["MISSING_ENV_VAR"].to_i` will return 0, which means our default option is never used.

We could put in some code that first tests for the existance, but we felt its cleaner and simpler to just make sure the env var is set!