Fix issue with dependency on env var in email
The email config in `config/environments/development.rb` was relying on checking on checking the value of `WCRS_FRONTEND_DOMAIN` directly, whereas we now set it in the `config/application.rb` (and default it if the env var is not set).

As we have a default, we tested in an environment where the env var was not set and immediately got an error on startup. This change brings the config inline by get `development.rb` to use the config value not the env var directly.