Resolve issues with config found in testing (#139)
https://eaflood.atlassian.net/browse/WC-281

Following changes made to bring the renewals service env vars and config inline with all other waste carrier registration apps some issues were found.

These changes resolve the issues found in this project.

Specifically they were

* Remove default values for registration charges

Having tried not setting these env vars and expecting the defaults to kick in, we realised that we had misunderstood how a command would work.

If an env var does not exist the call `ENV["MISSING_ENV_VAR"].to_i` will return 0, which means our default option is never used.

We could put in some code that first tests for the existance, but we felt its cleaner and simpler to just make sure the env var is set!

* Fix issue with dependency on env var in email

The email config in `config/environments/development.rb` was relying on checking on checking the value of `WCRS_FRONTEND_DOMAIN` directly, whereas we now set it in the `config/application.rb` (and default it if the env var is not set).

As we have a default, we tested in an environment where the env var was not set and immediately got an error on startup. This change brings the config inline by get `development.rb` to use the config value not the env var directly.

* Remove thin webserver

Removed the thin webserver (so now defaults to webrick) to bring the project inline with the others in the service.