Update to support using mock worldpay service
https://eaflood.atlassian.net/browse/WC-365

Though we have got the renewals service working with the new mock worldpay feature in the waste carriers service, the frontend is erroring because it makes http requests in a different way, specifically it assumes its always talking to a service using https.

The changes included here are to allow the frontend to talk to the mock when we need it to.

It also removes some obvious duplication.