Refactor mongoid config to use URI's (#141)
Previously we had specified each of the parameters that make up a connection to the databases within the `mongoid.yml` file. The documentation for the file is heavily centred on this kind of way of doing it <https://docs.mongodb.com/mongoid/master/tutorials/mongoid-installation/#anatomy-of-a-mongoid-config>.

So database, hosts, user, password etc are all listed out. However commented out is mention of providing a URI which has all the details included. Our web-ops team were pushing for using this format and after some investigation

- it looks like this is how we currently do it for our existing app when running in production
- though not scientific, some investigation online suggests this is the typical way to do it when running in production

So this change is about following conventions and switching the app to follow convention and use the URI. The one decision we have made is rather than maintaining both a URI and bunch of env vars that just duplicate the same info, we will use the URI format in all environments.

N.B. This change also deletes the options `max_retries` and `retry_interval` from the test config. Previous notes indicate this helped speed up tests however we cannot find any reference to these in the  existing Mongoid and MongoDb ruby driver documentation. Hence we believe they are just a left over from when Mongoid used Moped as its driver, and we just copied them across from the Frontend config.