Move renewal root path to logic not env var
We previously had made a change to our what we expected the env var WCRS_RENEWALS_DOMAIN to hold from `http://localhost` to `http:/localhost/fo` (for example).

However on reflection this is not reflected in the name of the env var, and our logic in the frontend is already setup to add a path to whatever value is held in it.

So this change amends the code based on us reverting our decision on what the env var will hold, so now the root path `/fo` will be included in the frontend logic.