Simplify the setting of urls in application.rb
Previously we were passing some config for the setting of URL's into a specicial method that chose whether to use the value from an env var if found, else the default value passed in, and then ensure the end result had http prefixed to it.

We're in the middle of trying to diagnose an issue with 2 assets not coming through because they are always reverting to using the default url of http://localhost:3000, and we feel it will make things easier if the setting of the urls' used was just the standard use env var else use default syntax.

Hence this method removes the helper method and tidies up the setting of the URL's to be consistent.