Remove hard coded use of service email values (#154)
When we send emails from non-production environments we are no longer using the production sendgrid details. As such this means if we were to continue to use 'registrations@wastecarriersregistration.service.gov.uk' as our from address it would be rejected by email clients such as Gmail because it does not match the DMARC policy set against it.

Therefore we have to specify a different email address. However in testing we found issues and realised that though we had an env var for this kind of information, the production details were still hardcoded into the Devise initializer.

This change removes it, and a no longer useful comment.