Add rake task to purge both databases (#156)
https://eaflood.atlassian.net/browse/WC-449

To allow us to run our acceptance tests against an environment and get reproducible results we need to be able to reset and re-seed the databases before each run.

Now we can do this locally using the rake commands and database scripts, but this is not so easy in our actual environments as you first need to be able to ssh in to the back end servers.

One of the rake commands is `mongoid:purge` however after checking it out it only purges the default client (in our case waste-carriers). So to support this goal we have added a new rake task that allows us to purge both databases.