Update renewal hold fix to not mess with env var
https://eaflood.atlassian.net/browse/WC-456

The previous version of this fix [PR 158](https://github.com/DEFRA/waste-carriers-frontend/pull/158) required us to amend the value held in WCRS_RENEWALS_DOMAIN. However realised after the fact this could effect the other apps.

So this update gives us a solution that does not require messing with the env var WCRS_RENEWALS_DOMAIN. as it currently is in production.