Fix missing GOV.UK logo in emails (#161)
https://eaflood.atlassian.net/browse/WC-458

We have spotted that the GOV.UK logo which makes up part of the header in emails is not shown. This needs to display to help give users confidence that the email is from a reputable source.

We believe https://github.com/DEFRA/waste-carriers-frontend/pull/152 is to blame for why it is no longer appearing, but we also know just adding back in the deleted code will cause issues elsewhere.

Hence this change copies the way Waste Exemptions includes images in emails, by instead adding it as an attachment and having the email link to that.