Fix broken renew link in back office
https://eaflood.atlassian.net/browse/WC-461

PR #160 dealt with an issue with the renewal links we spotted when app was deployed to our production environments.

However we have since spotted that the link displayed in the back office is using the public domain and hence is directing people to https://publicdomain/bo/renew/CBDU1234

Because of this we are getting a 404, because you cannot access the back office from the public domain.

This change deals with the issue.

WIP