Add CHANGELOG to the project
https://eaflood.atlassian.net/browse/WC-462

We ensure we make all changes to our projects using Pull Requests, and follow the standard for commits.

On that basis we use https://github.com/github-changelog-generator/github-changelog-generator as a way of automatically generating our CHANGELOG.

So this change covers adding the Gem and a rake task to simplify generating/updating the CHANGELOG.

We want a changelog because once we have put a project into production we need a way of cataloguing the changes we make to simplify the RfC process for getting them into production.