Remove change account email from admin screen
https://eaflood.atlassian.net/browse/WC-465

It has been agreed as a feature the change account email function is potentially risky, as it would have the ability to move all registrations to a new account, something users might not be aware of based on the current design.

As it is, it seems to be just causing problems due to its ability to orphan registrations if the email contains capital letters or the account is linked to more than one registration.

Therefore in a meeting on Sep 14 Chris Jackson the team lead agreed with the team that the link should be removed.