Display account_email for registrations in NCCC dashboard (#165)
https://eaflood.atlassian.net/browse/WC-477

When we list registrations for logged-in agency users, we should also include the account_email. This means that a backend user can quickly tell who a registration belongs to, and therefore what email address must be used to log in and renew or otherwise modify it.

We also frequently encounter instances where `account_email` is populated but there is no matching user record. These lead to support calls of which the first step by the delivery team is to ascertain whether the user record exists. Displaying the status of the user account next to the new account email field means both NCCC and the delivery team can see if this is the cause of an issue, e.g. the user knows their account email but cannot log in.