Update business types in locales to add missing
https://eaflood.atlassian.net/browse/WC-511

We noticed in [waste-carriers-back-office](https://github.com/DEFRA/waste-carriers-back-office) that we were missing the translation for overseas companies and [resolved that](https://github.com/DEFRA/waste-carriers-back-office/pull/145).

However our QA then spotted you would get a similiar error if you tried to edit a registration which had a business type of `overseas`.

So this fix updates the business type locales to match what is now in place in the new apps, to avoid missing translation errors and to keep things consistent.