Remove ref. to access code when init. from IR reg.
When we create a registration from an IR registration we specified that when doing so the reg_uuid and accessCode should not be overwritten. As we the access code is not actually used this doesn't matter, plus we don't have any more IR renewals to come through so the code is redundant anyway.