Update default renewal window to match current
When the service was updated and re-shipped on September 12 2018 we also updated the renewal window to 3 months instead of 6. This was in preparation for launch of the renewals service on the 19th.

However spotted that the default was still at 6 months.

This change updates the default to match what we now use, and because of this change we can remove an unnecessary stub.