Expand expired reg context to cover status & date
Registrations are marked as EXPIRED by the waste-carriers-service. We know the job runs at 8pm each day, and has been fixed to ignore the time element. However just to be sure, for example in the event the service goes down, the class checks both the status and the expires_on field.

So this change expands the context to cover both.