Add grace window config plkus rspec helpers
Add a new configuration value which can be driven by an env var (but has a default of 3 days which matches the current instruction from NCCC) for the grace window.

Also start to use this value in our tests, along with some helpers to determine a date that is both inside and outside of it.

(cherry picked from commit 2d22f4c9a40e4ef209cb275021f85b67f2fae849)