Implemented grace window into CanRenew con concern
Added the call into the `CanRenew.expired?` to `ExpiryDateService.in_expiry_grace_window?` to handle registrations within the "grace window".

Essentially realised we no longer care about the status of a registration when it comes to this logic, as it is all dependent on the value in expires_on. Hence also update the tests to reflect this (removing anything where the test is focused on just the status).

(cherry picked from commit 407780a6f4cad45cb8b8cd08964de380c8be3b3b)