Update CanBeRenewed.expired? concern method
Previous refactorings of the `expired?` method had been done in the context of ascertaining whether a registration could be renewed. However as a concern that applies to all registrations, what it should have been focused on was ascertaining whether the registration is expired. A registration that is in the 'grace window' is still expired and this method should reflect that.

Hence this change updates the tests and code to focus on determining only if a registration is expired.

N.B. This will break other tests that rely on checking whether an expired registration is in the grace window, but additional changes will be forthcoming.