Update can_renew? to include grace window check
Decided to go with including the grace window check in the can_renew? method. So now it only checks the grace window if a registration is `expired?`.

If this was under CodeClimate we would be getting issues regards the size of this method, but I think its acceptable for now as trying to split it up into individual methods in a class that is already massive would be more detrimental.