Add standard rubocop config to project
We know that the project does not follow our current rubocop style conventions. However when working on changes we want those changes to abide by them.

Most editors have some form of inbuilt linting, so by adding this config those editors can take advantage of the config being there to highlight issues in the code being worked on.

We as the devs than can use this information to ensure what we add or change meets our current conventions. It also means where an editor doesn't have a config file to fall back on, its doesn't cause confusion by highlighting issues that are not consistent with our conventions.